// app/features/indicators/language/types.ts
export type SourceLocation = {
  line: number
  column: number
  start: number
  end: number
}

export type FinScriptDiagnostic = SourceLocation & {
  severity: "error" | "warning"
  message: string
  code: string
}

export type InputDescriptor = {
  id: string
  title: string
  type: "int" | "float" | "bool" | "color"
  defaultValue: number | boolean | string
}